feat(app): Add delete button to rounds page
Added delete button to /rounds page in each row of the table that's generated. Now when you click on that button, it'll delete the round from the database and redirect you back to the /rounds page after reloading it. Fixed some formatting stuff in a couple of other HTML templates as well with the PyCharm Reformat Code function.